+++ Client/Service/CommentService.cs
@@ -10,6 +10,16 @@ public class CommentService : ICommentService
     private readonly ILogger<CommentService> _logger;
     private readonly Commenter.CommenterClient _client;
 
+    public List<Comment> StubComments = new List<Comment>
+    {
+        new() { Id = 1, PhotoId = 1, Title = "Test 1", Content = "Content goes here.." },
+        new() { Id = 2, PhotoId = 1, Title = "Test 2", Content = "Content goes here.." },
+        new() { Id = 3, PhotoId = 1, Title = "Test 3", Content = "Content goes here.." },
+        new() { Id = 4, PhotoId = 1, Title = "Test 4", Content = "Content goes here.." },
+        new() { Id = 5, PhotoId = 1, Title = "Test 5", Content = "Content goes here.." },
+        new() { Id = 6, PhotoId = 1, Title = "Test 6", Content = "Content goes here.." },
+    };
+
     public CommentService(ILogger<CommentService> logger, Commenter.CommenterClient client)
     {
         _logger = logger;
@@ -20,6 +30,12 @@ public class CommentService : ICommentService
     {
         try
         {
+            StubComments.Add(newComment);
+
+            await Task.CompletedTask;
+
+            return newComment;
+
             var request = new CreateRequest()
             {
                 Photoid = newComment.PhotoId,
@@ -47,6 +63,20 @@ public class CommentService : ICommentService
 
     public async Task<Comment?> DeleteAsync(int commentId)
     {
+        if (StubComments != null)
+            Console.WriteLine("stub is not null");
+
+        var item = StubComments?.FirstOrDefault(c => c.Id == commentId);
+        if (item != null)
+        {
+            StubComments.Remove(item);
+            return item;
+        }
+        else
+        {
+            return null;
+        }
+
         var request = new DeleteRequest() { Commentid = commentId };
         var response = await _client.DeleteAsync(request);
         return new()
@@ -63,6 +93,8 @@ public class CommentService : ICommentService
     {
         try
         {
+            return StubComments.Where(c => c.PhotoId == photoId).ToList();
+
             var request = new GetByPhotoIdRequest() { Photoid = photoId };
             var response = await _client.GetByPhotoIdAsync(request);
 
@@ -84,6 +116,22 @@ public class CommentService : ICommentService
 
     public async Task<Comment?> UpdateAsync(Comment changedComment)
     {
+        if (changedComment == null) return null;
+
+        var item = StubComments.FirstOrDefault(c => c.Id == changedComment.Id);
+
+        if (item != null)
+        {
+            StubComments.Remove(item);
+            StubComments.Add(item);
+
+            return changedComment;
+        }
+        else
+        {
+            return changedComment;
+        }
+
         var request = new UpdateRequest()
         {
             Id = changedComment.Id,
